requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
httpie==3.2.2
idna==3.7
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pillow==10.4.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.2.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
